+++ src/css/index.scss
@@ -13,7 +13,7 @@
 @import './preloader.scss';
 @import '../../src/components/forms/index.scss';
 @import '../../src/components/popup/popup.scss';
-@import '../pages/games/audiocall.scss';
+@import '../pages/games/audiocall/index.scss';
 
 @import './authorization.scss';
 @import './statistics/shortStatistics.scss';
