+++ src/utils/index.js
@@ -1,0 +1,5 @@
+export { requestCreator} from './requests'
+
+export function getRandomInt(max) {
+    return Math.floor(Math.random() * Math.floor(max));
+}
